feat: bundle optimization via code splitting & lazy load
