+++ package.json
@@ -20,8 +20,8 @@
     "client:test": "npm run test:vitest --workspace=@music-app/client"
   },
   "devDependencies": {
-    "@eslint/js": "^9.22.0",
-    "npm": "^10.0.0"
+    "@eslint/js": "^9.30.1",
+    "npm": "^11.4.2"
   },
   "engines": {
     "node": ">=18.0.0",
