+++ packages/client/package.json
@@ -2,72 +2,71 @@
   "dependencies": {
     "@connectrpc/connect-query": "^2.1.0",
     "@connectrpc/connect-web": "^2.0.2",
-    "@hookform/resolvers": "^5.0.1",
+    "@hookform/resolvers": "^5.1.1",
     "@mobily/ts-belt": "^3.13.1",
     "@music-app/proto": "*",
-    "@radix-ui/react-alert-dialog": "^1.1.10",
-    "@radix-ui/react-checkbox": "^1.2.2",
-    "@radix-ui/react-dialog": "^1.1.11",
-    "@radix-ui/react-dropdown-menu": "^2.1.11",
-    "@radix-ui/react-label": "^2.1.4",
-    "@radix-ui/react-popover": "^1.1.10",
-    "@radix-ui/react-select": "^2.2.2",
-    "@radix-ui/react-slider": "^1.3.2",
-    "@radix-ui/react-slot": "^1.2.0",
-    "@tailwindcss/vite": "^4.1.4",
-    "@tanstack/react-query": "^5.74.4",
+    "@radix-ui/react-alert-dialog": "^1.1.14",
+    "@radix-ui/react-checkbox": "^1.3.2",
+    "@radix-ui/react-dialog": "^1.1.14",
+    "@radix-ui/react-dropdown-menu": "^2.1.15",
+    "@radix-ui/react-label": "^2.1.7",
+    "@radix-ui/react-popover": "^1.1.14",
+    "@radix-ui/react-select": "^2.2.5",
+    "@radix-ui/react-slider": "^1.3.5",
+    "@radix-ui/react-slot": "^1.2.3",
+    "@tailwindcss/vite": "^4.1.11",
+    "@tanstack/react-query": "^5.81.5",
     "@tanstack/react-table": "^8.21.3",
-    "@vitejs/plugin-react": "^4.4.1",
-    "@wavesurfer/react": "^1.0.11",
+    "@vitejs/plugin-react": "^4.6.0",
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
     "cmdk": "^1.1.1",
-    "lucide-react": "^0.501.0",
+    "lucide-react": "^0.525.0",
     "next-themes": "^0.4.6",
     "react": "^19.1.0",
     "react-dom": "^19.1.0",
-    "react-error-boundary": "^5.0.0",
-    "react-hook-form": "^7.56.0",
-    "sonner": "^2.0.3",
-    "tailwind-merge": "^3.2.0",
-    "tailwindcss": "^4.1.4",
-    "wavesurfer.js": "^7.9.4",
-    "zod": "^3.24.3",
-    "zustand": "^5.0.3"
+    "react-error-boundary": "^6.0.0",
+    "react-hook-form": "^7.59.0",
+    "sonner": "^2.0.5",
+    "tailwind-merge": "^3.3.1",
+    "tailwindcss": "^4.1.11",
+    "wavesurfer.js": "^7.9.9",
+    "zod": "^3.25.69",
+    "zustand": "^5.0.6"
   },
   "devDependencies": {
     "@connectrpc/connect-playwright": "^0.6.0",
-    "@eslint/js": "^9.22.0",
-    "@faker-js/faker": "^9.8.0",
-    "@playwright/experimental-ct-react": "^1.53.1",
-    "@playwright/test": "^1.53.1",
-    "@testing-library/jest-dom": "^6.6.3",
+    "@eslint/js": "^9.30.1",
+    "@faker-js/faker": "^9.9.0",
+    "@playwright/experimental-ct-react": "^1.53.2",
+    "@playwright/test": "^1.53.2",
     "@testing-library/react": "^16.3.0",
-    "@testing-library/user-event": "^14.6.1",
-    "@types/node": "^22.14.1",
-    "@types/react": "^19.1.8",
-    "@types/react-dom": "^19.0.4",
-    "@vitejs/plugin-react-swc": "^3.8.0",
-    "eslint": "^9.22.0",
-    "eslint-import-resolver-typescript": "^4.3.4",
-    "eslint-plugin-import": "^2.31.0",
+    "@types/node": "^24.0.10",
+    "@types/react": "^18.3.23",
+    "@types/react-dom": "^18.3.7",
+    "@vitejs/plugin-react-swc": "^3.10.2",
+    "eslint": "^9.30.1",
+    "eslint-import-resolver-typescript": "^4.4.4",
+    "eslint-plugin-import": "^2.32.0",
     "eslint-plugin-react": "^7.37.5",
     "eslint-plugin-react-hooks": "^5.2.0",
-    "eslint-plugin-react-refresh": "^0.4.19",
+    "eslint-plugin-react-refresh": "^0.4.20",
     "eslint-plugin-vitest-globals": "^1.5.0",
-    "globals": "^16.0.0",
+    "globals": "^16.3.0",
     "jsdom": "^26.1.0",
-    "playwright": "^1.53.1",
-    "tw-animate-css": "^1.2.7",
-    "typescript": "~5.7.2",
-    "typescript-eslint": "^8.26.1",
-    "vite": "^6.3.1",
+    "playwright": "^1.53.2",
+    "rollup-plugin-visualizer": "^6.0.3",
+    "tw-animate-css": "^1.3.4",
+    "typescript": "~5.8.3",
+    "typescript-eslint": "^8.35.1",
+    "vite": "^7.0.0",
     "vitest": "^3.2.4"
   },
   "name": "@music-app/client",
   "private": true,
   "scripts": {
     "build": "tsc -b && vite build",
+    "build:analyze": "ANALYZE=true npm run build",
     "dev": "vite",
     "dev:with-proto": "npm run proto:generate:build && npm run dev",
     "generate:api": "openapi-zod-client \"${OPENAPI_URL:-http://localhost:8000/documentation/yaml}\" -o \"./src/generated/api.ts\" --export-schemas",
