+++ packages/client/src/components/TracksTable/TrackActions.tsx
@@ -1,18 +1,9 @@
 import { type Track } from "@music-app/proto/tracks";
 import { Edit, Ellipsis, Trash2, Upload } from "lucide-react";
-import { useRef, useState } from "react";
-import { toast } from "sonner";
+import { lazy, Suspense, useState } from "react";
 
-import DeleteTrackDialog from "@/components/TrackDialogs/DeleteTrackDialog";
-import EditTrackDialog from "@/components/TrackDialogs/EditTrackDialog";
 import { Button } from "@/components/ui/Button";
-import {
-  Dialog,
-  DialogContent,
-  DialogFooter,
-  DialogHeader,
-  DialogTitle,
-} from "@/components/ui/Dialog";
+import { DialogFallback } from "@/components/ui/DialogFallback";
 import {
   DropdownMenu,
   DropdownMenuContent,
@@ -20,154 +11,25 @@ import {
   DropdownMenuSeparator,
   DropdownMenuTrigger,
 } from "@/components/ui/DropdownMenu";
-import { Input } from "@/components/ui/Input";
-import { useUploadTrackFile } from "@/lib/api/tracks";
-import { cn } from "@/lib/utils";
+
+const DeleteTrackDialog = lazy(
+  () => import("@/components/TrackDialogs/DeleteTrackDialog")
+);
+const EditTrackDialog = lazy(
+  () => import("@/components/TrackDialogs/EditTrackDialog")
+);
+const UploadTrackDialog = lazy(
+  () => import("@/components/TrackDialogs/UploadTrackDialog")
+);
 
 interface TrackActionsProps {
   track: Track;
 }
 
-function UploadTrackDialog({
-  track,
-  open,
-  onOpenChange,
-}: {
-  track: Track;
-  open: boolean;
-  onOpenChange: (open: boolean) => void;
-}) {
-  const [file, setFile] = useState<File | null>(null);
-  const [isDragging, setIsDragging] = useState(false);
-
-  const uploadMutation = useUploadTrackFile();
-
-  const handleFileChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-    if (e.target.files?.[0]) {
-      setFile(e.target.files[0]);
-    }
-  };
-
-  const handleDrop = (e: React.DragEvent<HTMLDivElement>) => {
-    e.preventDefault();
-    e.stopPropagation();
-    setIsDragging(false);
-    if (e.dataTransfer.files[0]) {
-      setFile(e.dataTransfer.files[0]);
-    }
-  };
-
-  const handleDragOver = (e: React.DragEvent<HTMLDivElement>) => {
-    e.preventDefault();
-    e.stopPropagation();
-    setIsDragging(true);
-  };
-
-  const handleDragLeave = (e: React.DragEvent<HTMLDivElement>) => {
-    e.preventDefault();
-    e.stopPropagation();
-    setIsDragging(false);
-  };
-
-  const fileInputRef = useRef<HTMLInputElement>(null);
-  const handleDropzoneClick = () => {
-    fileInputRef.current?.click();
-  };
-
-  const handleUpload = async () => {
-    if (!file) return;
-
-    uploadMutation.mutate(
-      {
-        trackId: track.id,
-        filename: file.name,
-        content: new Uint8Array(await file.arrayBuffer()),
-        contentType: file.type,
-      },
-      {
-        onSuccess: () => {
-          toast.success("Audio file uploaded successfully!");
-          onOpenChange(false);
-          setFile(null);
-        },
-        onError: (error: Error) => {
-          toast.error(error.message || "Failed to upload file");
-        },
-      }
-    );
-  };
-
-  return (
-    <Dialog open={open} onOpenChange={onOpenChange}>
-      <DialogContent>
-        <DialogHeader>
-          <DialogTitle>
-            {track.audioFile ? "Replace Audio File" : "Upload Audio File"}
-          </DialogTitle>
-        </DialogHeader>
-
-        <Input
-          accept="audio/*"
-          className="hidden"
-          ref={fileInputRef}
-          type="file"
-          onChange={handleFileChange}
-        />
-
-        <div
-          className={cn(
-            "flex flex-col items-center justify-center p-4 border-2 border-dashed rounded-md cursor-pointer",
-            isDragging ? "border-primary bg-primary/10" : "border-border"
-          )}
-          style={{ minHeight: "100px" }}
-          onClick={handleDropzoneClick}
-          onDragLeave={handleDragLeave}
-          onDragOver={handleDragOver}
-          onDrop={handleDrop}
-        >
-          <Upload className="w-6 h-6 text-muted-foreground mb-1" />
-          <p className="text-sm text-muted-foreground">
-            {isDragging
-              ? "Drop the audio file here"
-              : "Drag & drop an audio file here, or click to select"}
-          </p>
-          {file && (
-            <p className="mt-2 text-sm font-medium text-primary">
-              Selected file: {file.name}
-            </p>
-          )}
-        </div>
-        <DialogFooter>
-          <Button
-            disabled={uploadMutation.isPending}
-            variant="outline"
-            onClick={() => {
-              onOpenChange(false);
-            }}
-          >
-            Cancel
-          </Button>
-          <Button
-            disabled={!file || uploadMutation.isPending}
-            onClick={() => {
-              void handleUpload();
-            }}
-          >
-            {uploadMutation.isPending
-              ? "Uploading..."
-              : track.audioFile
-              ? "Replace"
-              : "Upload"}
-          </Button>
-        </DialogFooter>
-      </DialogContent>
-    </Dialog>
-  );
-}
-
 export function TrackActions({ track }: TrackActionsProps) {
   const [isEditDialogOpen, setIsEditDialogOpen] = useState(false);
   const [isUploadDialogOpen, setIsUploadDialogOpen] = useState(false);
+  const [isDeleteDialogOpen, setIsDeleteDialogOpen] = useState(false);
 
   return (
     <>
@@ -203,32 +65,51 @@ export function TrackActions({ track }: TrackActionsProps) {
             <span>{track.audioFile ? "Replace" : "Upload"} File</span>
           </DropdownMenuItem>
           <DropdownMenuSeparator />
-          <DeleteTrackDialog track={track}>
-            <DropdownMenuItem
-              className="text-destructive focus:text-destructive focus:bg-destructive/10"
-              data-testid={`delete-track-${track.id}`}
-              onSelect={(e) => {
-                e.preventDefault();
-              }}
-            >
-              <Trash2 className="mr-2 h-4 w-4" />
-              <span>Delete Track</span>
-            </DropdownMenuItem>
-          </DeleteTrackDialog>
+          <DropdownMenuItem
+            className="text-destructive focus:text-destructive focus:bg-destructive/10"
+            data-testid={`delete-track-${track.id}`}
+            onClick={() => {
+              setIsDeleteDialogOpen(true);
+            }}
+          >
+            <Trash2 className="mr-2 h-4 w-4" />
+            <span>Delete Track</span>
+          </DropdownMenuItem>
         </DropdownMenuContent>
       </DropdownMenu>
 
-      <EditTrackDialog
-        open={isEditDialogOpen}
-        track={track}
-        onOpenChange={setIsEditDialogOpen}
-      />
+      {isEditDialogOpen && (
+        <Suspense fallback={<DialogFallback />}>
+          <EditTrackDialog
+            open={isEditDialogOpen}
+            track={track}
+            onOpenChange={setIsEditDialogOpen}
+          />
+        </Suspense>
+      )}
 
-      <UploadTrackDialog
-        open={isUploadDialogOpen}
-        track={track}
-        onOpenChange={setIsUploadDialogOpen}
-      />
+      {isUploadDialogOpen && (
+        <Suspense fallback={<DialogFallback />}>
+          <UploadTrackDialog
+            open={isUploadDialogOpen}
+            track={track}
+            onOpenChange={setIsUploadDialogOpen}
+          />
+        </Suspense>
+      )}
+
+      {isDeleteDialogOpen && (
+        <Suspense fallback={<DialogFallback />}>
+          <DeleteTrackDialog
+            open={isDeleteDialogOpen}
+            track={track}
+            onDialogClose={() => {
+              setIsDeleteDialogOpen(false);
+            }}
+            onOpenChange={setIsDeleteDialogOpen}
+          />
+        </Suspense>
+      )}
     </>
   );
 }
