+++ packages/client/src/components/ui/DialogFallback.tsx
@@ -1,0 +1,5 @@
+export const DialogFallback = () => (
+  <div className="fixed inset-0 z-50 bg-black/60 flex items-center justify-center">
+    <div className="animate-spin h-6 w-6 border-2 border-white border-t-transparent rounded-full"></div>
+  </div>
+);
